+++ CHANGELOG.md
@@ -15,6 +15,9 @@ for further information about branching and tagging conventions.
 #### Changed
 - Now using @tokend/js-sdk@1.7.0-rc.2
 
+#### Fixed
+- Pagination while loading limits on limits management page
+
 ## [1.7.0-rc.2] - 2019-05-21
 ### "Under the hood" changes
 #### Changed